refactor(tools): move pipeline policy args out of resolver
